# depot-lifecycle-openapi-2.2.5.yaml
openapi: 3.0.1
info:
  title: Depot Life Cycle
  description: "# Purpose\n\n A depot centric API for managing the interchange activity\
    \ & repair lifecycle of a shipping container.  The API is expected to be used\
    \ by Customers, Depots, and Owners to facilitate real time communication between\
    \ systems instead of traditional EDI files.\n\n\n\n # Overview\n\n The depot lifecycle\
    \ API is a RESTful API.  It's requests & responses are loosely based upon traditional\
    \ EDI files.  For code definitions, explanations, or traditional EDI definitions\
    \ refer to [IICL TB 002, February 2003](https://www.iicl.org/iiclforms/assets/File/public/bulletins/TB002_EDIS_February_2003.pdf).\n\
    \n\n\n # JSON\n\n The requests and responses of this API are formatted according\
    \ to the [JSON](https://www.json.org/) standard.\n\n ## Null vs Absent\n\n For\
    \ all types other than arrays, nullable and absent are equivalent by this specification.\
    \  For example, if a non-array attribute is marked optional it may be excluded\
    \ from the JSON or it may be explicitly set to null.  However, to keep request\
    \ sizes small, it is recommended to prefer excluding attributes rather than sending\
    \ an explicit null.\n\n For arrays, null should be avoided.  For example, if the\
    \ array is empty or there are no values expected for the array then either the\
    \ field can be excluded from the JSON or an empty array should be transmitted.\n\
...
  contact:
...
servers:
...
security:
- Dynamic_Token: []
- Static_Token: []
...
paths:
...
  /api/v2/workOrder/{workOrderNumber}:
    put:
      tags:
      - workOrder
      summary: updates a repair work order
      description: Submits updates to the given work order.
      operationId: updateWorkOrder
      parameters:
      - name: workOrderNumber
        in: path
        description: the work order number
        required: true
        schema:
          maxLength: 16
          type: string
          example: WHAMG30001
      requestBody:
        description: the updated work order record
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrder"
        required: true
      responses:
        "200":
          description: successfully update the workOrder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
        "400":
          description: an error occurred trying to update the work order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: work order updates are not allowed by security
        "404":
          description: "shipping container, depot, or work order could not be found"
        "501":
          description: this feature is not supported by this server
        "503":
          description: "API is temporarily paused, and not accepting any activity"
      security:
      - JWT_TOKEN: []
  /api/v2/workOrderUnit/{workOrderNumber}:
    put:
      tags:
      - workOrder
      summary: marks a shipping container repaired
      description: "For the given work order, attempt to mark it repair complete."
      operationId: updateWorkOrderUnit
      parameters:
      - name: workOrderNumber
        in: path
        description: the work order number
        required: true
        schema:
          maxLength: 16
          type: string
          example: WHAMG30001
      requestBody:
        description: Necessary information to mark a shipping container repair complete
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepairComplete"
        required: true
      responses:
        "200":
          description: successfully repair completed the workOrder
          content:
            application/json:
              schema: {}
        "400":
          description: an error occurred trying to repair complete the work order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: repair completion is not allowed by security
        "404":
          description: "shipping container, depot, or work order could not be found"
        "501":
          description: this feature is not supported by this server
        "503":
          description: "API is temporarily paused, and not accepting any activity"
      security:
      - JWT_TOKEN: []
      x-iicl-purpose: activity
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          maxLength: 6
          pattern: "^[A-Z0-9]{3}[0-9]{3}$"
          required:
          - "false"
          type: string
          description: indicator code specific to this error
          nullable: true
          example: TRI521
        message:
          required:
          - "false"
          type: string
          description: a descriptive error message
          nullable: true
          example: Info TRI521 - Unit has been gated-in but is not off-hired
        details:
          required:
          - "false"
          type: array
          description: details for the descriptive error message
          nullable: false
          items:
            type: string
            example: "Customer may turn in unit, but will continued to be billed."
      description: |-
        a general response object used when an error occurs to describe why it occurred

        `400` http status codes are used when a request is invalid or the basic requirements are not met

        `405` http status codes are used for business validations
...
    Party:
      required:
      - companyId
      type: object
      properties:
        companyId:
          maxLength: 9
          pattern: "^[A-Z0-9]{9}$"
          required:
          - "true"
          type: string
          description: "the identifier for this party, often referred to as an EDI\
            \ Address"
          nullable: false
          example: DEHAMCMRA
        userCode:
          maxLength: 16
          required:
          - "false"
          type: string
          description: the user identifier at this company that concerns this message
          nullable: true
          example: JDOE
        userName:
          maxLength: 70
          required:
          - "false"
          type: string
          description: the full name for the user identified by `userCode`
          nullable: true
          example: John Doe
        faxNumber:
          required:
          - "false"
          type: array
          description: the contact fax number(s) for this party
          nullable: false
          items:
            type: string
        phoneNumber:
          required:
          - "false"
          type: array
          description: the contact phone number(s) for this party
          nullable: false
          items:
            type: string
        emailAddress:
          required:
          - "false"
          type: array
          description: the contact email address(es) for this party
          nullable: false
          items:
            type: string
        name:
          maxLength: 150
          required:
          - "false"
          type: string
          description: the name of this company
          nullable: true
          example: CMR Container Maintenance Rep.
        code:
          maxLength: 10
          required:
          - "false"
          type: string
          description: "the internal system code for this company, will be system\
            \ specific to the system delivering or receiving this message"
          nullable: true
          example: HAMG
        streetAddress:
          required:
          - "false"
          type: array
          description: the street address lines
          nullable: false
          items:
            type: string
        city:
          maxLength: 28
          required:
          - "false"
          type: string
          description: the city for the address
          nullable: true
        country:
          maxLength: 2
          required:
          - "false"
          type: string
          description: the 2 digit ISO country code for the address
          nullable: true
        postalCode:
          maxLength: 20
          required:
          - "false"
          type: string
          description: the postal code for the address
          nullable: true
        stateProvince:
          maxLength: 20
          required:
          - "false"
          type: string
          description: the optional state or province code for the address
          nullable: true
        latitude:
          required:
          - "false"
          type: number
          description: "Instead of an address, provides the latitude of this party"
          format: double
          nullable: true
        longitude:
          required:
          - "false"
          type: number
          description: "Instead of an address, provides the longitude of this party"
          format: double
          nullable: true
      description: represents a company (or location) involving shipping containers
...
    RepairComplete:
      required:
      - completionDate
      - depot
      - unitNumber
      - workOrderNumber
      type: object
      properties:
        workOrderNumber:
          maxLength: 16
          minLength: 1
          required:
          - "true"
          type: string
          description: the identifier of the work order to repair complete
          nullable: false
          example: WHAMG46019
        depot:
          description: the storage location where the shipping container is being
            repaired
          allOf:
          - $ref: "#/components/schemas/Party"
        completionDate:
          required:
          - "true"
          type: string
          description: |-
            the date and time in the depot local time zone (i.e. `2018-04-10T19:37:04Z`) that this work order was repaired.

            ( notation as defined by [RFC 3339, section 5.6](https://tools.ietf.org/html/rfc3339#section-5.6) )
          format: date-time
          nullable: false
        unitNumber:
          maxLength: 11
          pattern: "^[A-Z]{4}[X0-9]{6}[A-Z0-9]{0,1}$"
          required:
          - "true"
          type: string
          description: the unit number of the shipping container at the time of repair
            approval
          nullable: false
          example: CONU1234561
      description: Necessary information to mark a specific unit under a work order
        repaired
...
    WorkOrder:
      required:
      - approvalDate
      - depot
      - lineItems
      - owner
      - type
      - workOrderNumber
      type: object
      properties:
        relatedId:
          required:
          - "false"
          type: integer
          description: |-
            *Field is currently proposed to be added - not currently production approved.*

            An internal system identifier to be used to upload Estimate Photos or compare related activities.
          format: int64
          nullable: true
          example: 10102561
        workOrderNumber:
          maxLength: 16
          minLength: 1
          required:
          - "true"
          type: string
          description: the identifier for this work order; this will be the approval
            number for repairs
          nullable: false
          example: WHAMG46019
        depot:
          description: the storage location where the shipping container is being
            repaired
          allOf:
          - $ref: "#/components/schemas/Party"
        owner:
          description: the owner of the shipping container
          allOf:
          - $ref: "#/components/schemas/Party"
        billingParty:
          description: the party that will bill the customer portion of damages for
            this repair
          nullable: true
          allOf:
          - $ref: "#/components/schemas/Party"
        type:
          maxLength: 11
          required:
          - "true"
          type: string
          description: the type of repair approved
          nullable: false
          example: SELLCWCA
        approvalDate:
          required:
          - "true"
          type: string
          description: |-
            the date and time in the depot local time zone (i.e. `2017-05-10T19:37:04Z`) that this work order is authorized for repair

            ( notation as defined by [RFC 3339, section 5.6](https://tools.ietf.org/html/rfc3339#section-5.6) )
          format: date-time
          nullable: false
        approvalTotal:
          required:
          - "false"
          type: number
          description: the total amount approved for repair
          format: double
          nullable: true
          example: 175
        approvalCurrency:
          maxLength: 3
          pattern: "^[A-Z]{3}$"
          required:
          - "false"
          type: string
          description: the currency of the approval total
          nullable: true
          example: EUR
        expirationDate:
          required:
          - "false"
          type: string
          description: |-
            the date and time in the depot local time zone (i.e. `2020-07-21T17:32:28Z`) that this repair must be completed by

            ( notation as defined by [RFC 3339, section 5.6](https://tools.ietf.org/html/rfc3339#section-5.6) )
          format: date-time
          nullable: true
        comments:
          maxLength: 500
          required:
          - "false"
          type: string
          description: comments pertaining to this repair for the intended recipient
            of this message
          nullable: true
          example: CWCA repairs for unit TCKU3456654 total 175.00 EUR per DEHAMCE1856373.1
        lineItems:
          maxItems: 200
          minItems: 1
          required:
          - "true"
          type: array
          description: units associated to this work order
          nullable: false
          items:
            $ref: "#/components/schemas/WorkOrderUnit"
      description: An approval of a damage estimate signifying a depot may repair
        a shipping container
    WorkOrderUnit:
      required:
      - effectiveInspectionCriteria
      - status
      - unitNumber
      type: object
      properties:
        estimateNumber:
          maxLength: 16
          minLength: 1
          required:
          - "false"
          type: string
          description: The estimate number that initiated this unit on this work order
          nullable: true
          example: DEHAMCE1856373
        unitNumber:
          maxLength: 11
          pattern: "^[A-Z]{4}[X0-9]{6}[A-Z0-9]{0,1}$"
          required:
          - "true"
          type: string
          description: the unit number of the shipping container at the time of repair
            approval
          nullable: false
          example: CONU1234561
        effectiveInspectionCriteria:
          maxLength: 10
          required:
          - "true"
          type: string
          description: repair the shipping container to this grade or category standard
          nullable: false
          example: CWCA-1
        remark:
          maxLength: 11
          pattern: "^[A-Z]{4}[X0-9]{6}[A-Z0-9]{0,1}$"
          required:
          - "false"
          type: string
          description: the unit number to remark the shipping container on repair
          nullable: true
        releaseNumber:
          maxLength: 16
          required:
          - "false"
          type: string
          description: the Release approved for gate out after repair
          nullable: true
        status:
          description: |-
            an indicator of the unit state on this work order

            `TIED` - shipping container is considered under repair on this work order

            `REMOVED` - shipping container is removed from this work order

            `REPAIRED` - shipping container is considered repaired
          example: TIED
          allOf:
          - $ref: "#/components/schemas/WorkOrderUnitStatus"
      description: information for a specific unit on a work order
    WorkOrderUnitStatus:
      type: string
      description: |-
        an indicator of the unit state on this work order

        `TIED` - shipping container is considered under repair on this work order

        `REMOVED` - shipping container is removed from this work order

        `REPAIRED` - shipping container is considered repaired
      example: TIED
      enum:
      - TIED
      - REMOVED
      - REPAIRED
  securitySchemes:
    JWT_TOKEN:
      type: http
      description: JWT Bearer Authentication
      scheme: bearer
      bearerFormat: JWT
...